Started packaging the templates
