+++ src/main/java/github/jdrost1818/plaster/service/task/generate/ControllerItGenerate.java
@@ -1,0 +1,19 @@
+package github.jdrost1818.plaster.service.task.generate;
+
+import github.jdrost1818.plaster.service.ServiceProvider;
+import github.jdrost1818.plaster.service.task.PlasterTaskId;
+import github.jdrost1818.plaster.service.task.util.DoNothing;
+
+public class ControllerItGenerate extends GenerateTask {
+
+    public ControllerItGenerate() {
+        super(
+                "Could not generate controller it doc tests",
+                PlasterTaskId.CONTROLLER_IT,
+                ServiceProvider.getControllerTestTemplateService(),
+                new DoNothing()
+        );
+    }
+
+
+}
