+++ src/main/java/github/jdrost1818/plaster/service/task/generate/ControllerTestGenerate.java
@@ -1,7 +1,9 @@
 package github.jdrost1818.plaster.service.task.generate;
 
+import github.jdrost1818.plaster.data.Setting;
 import github.jdrost1818.plaster.service.ServiceProvider;
 import github.jdrost1818.plaster.service.task.PlasterTaskId;
+import github.jdrost1818.plaster.service.task.util.CheckSetting;
 import github.jdrost1818.plaster.service.task.util.DoNothing;
 
 public class ControllerTestGenerate extends GenerateTask {
@@ -11,7 +13,10 @@ public class ControllerTestGenerate extends GenerateTask {
                 "Could not generate controller tests",
                 PlasterTaskId.CONTROLLER_TEST,
                 ServiceProvider.getControllerTestTemplateService(),
-                new DoNothing()
+                new CheckSetting(
+                        Setting.IS_REST_DOCUMENTATION_TESTING_ENABLED,
+                        new ControllerITDocGenerate(),
+                        new ControllerItGenerate())
         );
     }
