+++ src/main/java/github/jdrost1818/plaster/service/template/repository/RepositoryTestTemplateService.java
@@ -1,8 +1,9 @@
-package github.jdrost1818.plaster.service.template;
+package github.jdrost1818.plaster.service.template.repository;
 
 import github.jdrost1818.plaster.data.TemplateType;
 import github.jdrost1818.plaster.domain.FileInformation;
 import github.jdrost1818.plaster.service.ConfigurationService;
+import github.jdrost1818.plaster.service.template.TemplateService;
 import org.jtwig.JtwigModel;
 import org.jtwig.JtwigTemplate;
 
@@ -13,12 +14,12 @@ public class RepositoryTestTemplateService extends TemplateService {
     }
 
     @Override
-    JtwigModel addCustomInformation(JtwigModel model, FileInformation fileInformation) {
+    protected JtwigModel addCustomInformation(JtwigModel model, FileInformation fileInformation) {
         return super.addTypeField(model, fileInformation.getClassName(), TemplateType.REPOSITORY_TEST);
     }
 
     @Override
-    JtwigTemplate getTemplate() {
+    protected JtwigTemplate getTemplate() {
         return JtwigTemplate.classpathTemplate("template/repository/repositoryTest.twig");
     }
