+++ src/test/java/github/jdrost1818/plaster/service/template/repository/RepositoryTemplateServiceTest.java
@@ -1,4 +1,4 @@
-package github.jdrost1818.plaster.service.template;
+package github.jdrost1818.plaster.service.template.repository;
 
 import com.google.common.collect.Lists;
 import github.jdrost1818.plaster.data.Setting;
@@ -7,6 +7,7 @@ import github.jdrost1818.plaster.data.TemplateType;
 import github.jdrost1818.plaster.domain.*;
 import github.jdrost1818.plaster.domain.template.FlattenedField;
 import github.jdrost1818.plaster.service.ConfigurationService;
+import github.jdrost1818.plaster.service.template.repository.RepositoryTemplateService;
 import org.jtwig.JtwigModel;
 import org.jtwig.JtwigTemplate;
 import org.junit.Before;
@@ -22,12 +23,12 @@ import static org.hamcrest.Matchers.hasSize;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
-public class ServiceTemplateServiceTest {
+public class RepositoryTemplateServiceTest {
 
     @Mock
     private ConfigurationService configurationService;
 
-    private ServiceTemplateService classUnderTest;
+    private RepositoryTemplateService classUnderTest;
 
     private FileInformation fileInformation;
 
@@ -35,7 +36,7 @@ public class ServiceTemplateServiceTest {
     public void setUp() throws Exception {
         initMocks(this);
 
-        this.classUnderTest = new ServiceTemplateService(configurationService);
+        this.classUnderTest = new RepositoryTemplateService(configurationService);
 
         Field id = new Field(new TypeDeclaration("List", Lists.newArrayList(StoredJavaType.LIST.getType(false))), "id");
         Field mapField = new Field(new TypeDeclaration("Map", Lists.newArrayList(StoredJavaType.MAP.getType(false))), "var1");
@@ -45,23 +46,6 @@ public class ServiceTemplateServiceTest {
         this.fileInformation = new FileInformation("example_class", id, Lists.newArrayList(
                 mapField, listField, exampleField
         ));
-    }
-
-    @Test
-    public void addTypeField() throws Exception {
-        JtwigModel model = JtwigModel.newModel();
-
-        when(this.configurationService.get(Setting.APP_PATH)).thenReturn("/com/example/app");
-        when(this.configurationService.get(Setting.REL_SERVICE_PACKAGE)).thenReturn("/service");
-        when(this.configurationService.get(Setting.SUB_DIR_PATH)).thenReturn("/somewhere");
-
-        JtwigModel modifiedModel = this.classUnderTest.addTypeField(model, "example_class", TemplateType.SERVICE);
-
-        FlattenedField x = (FlattenedField) modifiedModel.get("serviceField").get().getValue();
-
-        assertThat(x.getClassName(), equalTo("ExampleClassService"));
-        assertThat(x.getPackagePath(), equalTo("com.example.app.service.somewhere"));
-        assertThat(x.getVarName(), equalTo("exampleClassService"));
     }
 
     @Test
@@ -81,7 +65,6 @@ public class ServiceTemplateServiceTest {
 
         assertThat(dependencies, hasSize(1));
         assertThat(modifiedModel.get("modelField").get().getValue(), equalTo(new FlattenedField("", "ExampleClass", "exampleClass")));
-        assertThat(modifiedModel.get("serviceField").get().getValue(), equalTo(new FlattenedField("", "ExampleClassService", "exampleClassService")));
         assertThat(modifiedModel.get("repoField").get().getValue(), equalTo(new FlattenedField("", "ExampleClassRepository", "exampleClassRepository")));
         assertThat(modifiedModel.get("idField").get().getValue(), equalTo(new FlattenedField("", "List", "id")));
     }
@@ -89,41 +72,22 @@ public class ServiceTemplateServiceTest {
     @Test
     public void renderTemplate() {
         String expected = "" +
-                "package com.example.app.service.somewhere;\n" +
+                "package com.example.app.repository.somewhere;\n" +
                 "\n" +
-                "import org.springframework.beans.factory.annotation.Autowired;\n" +
-                "import org.springframework.stereotype.Service;\n" +
+                "import org.springframework.data.domain.Page;\n" +
+                "import org.springframework.data.domain.Pageable;\n" +
+                "import org.springframework.data.jpa.domain.Specification;\n" +
+                "import org.springframework.data.repository.CrudRepository;\n" +
                 "\n" +
                 "import java.util.List;\n" +
                 "\n" +
                 "import com.example.app.model.somewhere.ExampleClass;\n" +
-                "import com.example.app.repository.somewhere.ExampleClassRepository;\n" +
                 "\n" +
-                "@Service\n" +
-                "public class ExampleClassService {\n" +
+                "public interface ExampleClassRepository extends CrudRepository<ExampleClass, List> {\n" +
                 "\n" +
-                "    private final ExampleClassRepository exampleClassRepository;\n" +
-                "    \n" +
-                "    @Autowired\n" +
-                "    public ExampleClassService(ExampleClassRepository exampleClassRepository) {\n" +
-                "        this.exampleClassRepository = exampleClassRepository;\n" +
-                "    }\n" +
-                "    \n" +
-                "    public ExampleClass create(ExampleClass exampleClass) {\n" +
-                "        return this.exampleClassRepository.save(exampleClass);\n" +
-                "    }\n" +
-                "    \n" +
-                "    public ExampleClass read(List id) {\n" +
-                "        return this.exampleClassRepository.findOne(id);\n" +
-                "    }\n" +
-                "    \n" +
-                "    public ExampleClass update(ExampleClass exampleClass) {\n" +
-                "        return this.exampleClassRepository.save(exampleClass);\n" +
-                "    }\n" +
-                "    \n" +
-                "    public void delete(List id) {\n" +
-                "        this.exampleClassRepository.delete(id);\n" +
-                "    }\n" +
+                "    Page<ExampleClass> findAll(Specification<ExampleClass> spec, Pageable pageInfo);\n" +
+                "\n" +
+                "    ExampleClass findOne(Specification<ExampleClass> spec);\n" +
                 "\n" +
                 "}";
